Move heuritics for base canonical form into a separate file
